look/look/requirements.txt
black==23.11.0 #УБРАТЬ ЛИШНЕЕ!!!!!
cffi==1.16.0
click==8.1.7
cryptography==41.0.5
exceptiongroup==1.1.3
falcon==3.1.1
greenlet==3.0.0
gunicorn==21.2.0
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==23.2
passlib==1.7.4
pathspec==0.11.2
platformdirs==4.0.0
pluggy==1.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
pytest==7.4.2
SQLAlchemy==2.0.22
tomli==2.0.1
typing_extensions==4.8.0